Simplify ExplorationLauncher, extract api specific bits.
Also added ComponentComposition static methods to configure
explorations based on column type.
